chore: improve our project permissions label, move AccessOverviewAccordion into its folder
